// src/app/api-client/model/game.ts
/**
 * RAWG Video Games Database API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GameEsrbRating } from './gameEsrbRating';
import { GamePlatformsInner } from './gamePlatformsInner';


export interface Game { 
    readonly id?: number;
    readonly slug?: string;
    readonly name?: string;
    readonly released?: string;
    readonly tba?: boolean;
    readonly background_image?: string;
    rating: number;
    readonly rating_top?: number;
    readonly ratings?: object;
    readonly ratings_count?: number;
    readonly reviews_text_count?: string;
    readonly added?: number;
    readonly added_by_status?: object;
    readonly metacritic?: number;
    /**
     * in hours
     */
    readonly playtime?: number;
    readonly suggestions_count?: number;
    readonly updated?: string;
    esrb_rating?: GameEsrbRating | null;
    platforms?: Array<GamePlatformsInner>;
}