// src/app/api-client/model/gamesList200Response.ts
/**
 * RAWG Video Games Database API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Game } from './game';


export interface GamesList200Response { 
    count: number;
    next?: string | null;
    previous?: string | null;
    results: Array<Game>;
}