// src/app/api-client/model/genre.ts
/**
 * RAWG Video Games Database API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface Genre {
  readonly id?: number;
  name: string;
  readonly slug?: string;
  readonly games_count?: number;
  readonly image_background?: string;
}