// src/app/api-client/model/genresList200Response.ts
/**
 * RAWG Video Games Database API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Genre } from './genre';


export interface GenresList200Response { 
    count: number;
    next?: string | null;
    previous?: string | null;
    results: Array<Genre>;
}